+++ Group Project 1/graph.py
@@ -2,6 +2,8 @@
 # Majority of this code is pulled from the Artificial Intelligence: A Modern Approach reference code
 # Repo located at https://github.com/aimacode/aima-python
 
+from util import *
+
 
 class Graph:
     """A graph connects nodes (vertices) by edges (links). Each edge can also
@@ -61,6 +63,45 @@ class Graph:
 def UndirectedGraph(graph_dict=None):
     """Build a Graph where every edge (including future ones) goes both ways."""
     return Graph(graph_dict=graph_dict, directed=False)
+
+
+class GraphProblem(Problem):
+    """The problem of searching a graph from one node to another."""
+
+    def __init__(self, initial, goal, graph):
+        super().__init__(initial, goal)
+        self.graph = graph
+
+    def actions(self, A):
+        """The actions at a graph node are just its neighbors."""
+        return list(self.graph.get(A).keys())
+
+    def result(self, state, action):
+        """The result of going to a neighbor is just that neighbor."""
+        return action
+
+    def path_cost(self, cost_so_far, A, action, B):
+        return cost_so_far + (self.graph.get(A, B) or np.inf)
+
+    def find_min_edge(self):
+        """Find minimum value of edges."""
+        m = np.inf
+        for d in self.graph.graph_dict.values():
+            local_min = min(d.values())
+            m = min(m, local_min)
+
+        return m
+
+    def h(self, node):
+        """h function is straight-line distance from a node's state to goal."""
+        locs = getattr(self.graph, 'locations', None)
+        if locs:
+            if type(node) is str:
+                return int(distance(locs[node], locs[self.goal]))
+
+            return int(distance(locs[node.state], locs[self.goal]))
+        else:
+            return np.inf
 
 
 # Romania map represented as a graph.
